GCC: Switch from ceilf to ceil

+++ src/game/Entities/StatSystem.cpp
@@ -256,7 +256,7 @@ void Unit::UpdateMaxPower(Powers power)
     value += GetModifierValue(unitMod, TOTAL_VALUE) +  bonusPower;
     value *= GetModifierValue(unitMod, TOTAL_PCT);
 
-    SetMaxPower(power, uint32(std::ceilf(value)));
+    SetMaxPower(power, uint32(std::ceil(value)));
 }
 
 void Player::ApplyFeralAPBonus(int32 amount, bool apply)
@@ -1130,7 +1130,7 @@ void Pet::UpdateMaxPower(Powers power)
     value += GetModifierValue(unitMod, TOTAL_VALUE) + std::max((addValue - 20) * 15 + 20, 0.f);
     value *= GetModifierValue(unitMod, TOTAL_PCT);
 
-    SetMaxPower(power, uint32(std::ceilf(value)));
+    SetMaxPower(power, uint32(std::ceil(value)));
 }
 
 void Pet::UpdateAttackPowerAndDamage(bool ranged)
